# API/swagger.yaml
openapi: 3.0.0
info:
  title: FintechBankApp API
  description: API completa para o sistema bancário FintechBankApp
  version: 1.0.0
  contact:
    name: FintechBankApp Team
    email: support@fintechbankapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Servidor de desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operação realizada com sucesso"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erro na operação"
...
    RecurringBill:
      type: object
      description: |
        Conta recorrente normalizada (recurringBillsRepo.normalize).
        Os campos userFullName/userRole só existem na rota admin (join com users).
      properties:
        id:
          type: string
          example: "uuid-xxxx"
        cpf:
          type: string
          example: "11111111111"
        name:
          type: string
          example: "Netflix"
        amount:
          type: number
          format: float
          example: 29.90
        dueDay:
          type: integer
          example: 15
        category:
          type: string
          example: "cultura"
        frequency:
          type: string
          enum: [MONTHLY, ANNUAL]
          example: "MONTHLY"
        paymentMethod:
          type: string
          enum: [ACCOUNT_DEBIT, CREDIT_CARD]
          example: "ACCOUNT_DEBIT"
        status:
          type: string
          enum: [active, past_due, suspended, canceled, paid]
          example: "active"
        cardToken:
          type: string
          nullable: true
        planId:
          type: string
          nullable: true
        nextBillingDate:
          type: string
          format: date-time
          nullable: true
        retryCount:
          type: integer
          example: 0
        maxRetries:
          type: integer
          example: 3
        lastRetryAt:
          type: string
          format: date-time
          nullable: true
        failureReason:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userFullName:
          type: string
          nullable: true
          description: Nome do cliente (somente na rota GET /admin/recurring-bills)
        userRole:
          type: string
          nullable: true
          description: Role do cliente (somente na rota GET /admin/recurring-bills)
    RecurringBillRequest:
      type: object
      required: [name, amount, dueDay]
      properties:
        name:
          type: string
          example: "Netflix"
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 39.90
        dueDay:
          type: integer
          minimum: 1
          maximum: 31
          example: 15
        category:
          type: string
          example: "cultura"
    RecurringBillPayRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [ACCOUNT_DEBIT, CREDIT_CARD]
          default: ACCOUNT_DEBIT
          example: ACCOUNT_DEBIT
          description: >
            ACCOUNT_DEBIT debita users.balance (não toca a fatura);
            CREDIT_CARD consome credit_card_available_limit e grava
            PAYMENT com sufixo (Faturado no Cartão) — fora da fatura.
        name:
          type: string
          description: Obrigatório apenas no upsert (conta existe só no localStorage do frontend).
          example: "Netflix"
        amount:
          type: number
          format: float
          minimum: 0.01
          description: Obrigatório apenas no upsert.
          example: 29.90
        dueDay:
          type: integer
          minimum: 1
          maximum: 31
          description: 'Opcional no upsert (default: dia atual).'
          example: 7
        category:
          type: string
          description: 'Opcional no upsert (default: outros).'
          example: "cultura"
        frequency:
          type: string
          enum: [MONTHLY, ANNUAL]
          default: MONTHLY
          description: Opcional no upsert — define o avanço do ciclo (next_billing_date).
          example: MONTHLY
...
paths:
...
  /recurring-bills/{cpf}:
    get:
      tags: [Financeiro]
      summary: Listar contas recorrentes
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
      responses:
        '200':
          description: Lista de contas recorrentes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  bills:
                    type: array
                    items: { $ref: '#/components/schemas/RecurringBill' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Financeiro]
      summary: Criar conta recorrente
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecurringBillRequest' }
      responses:
        '201':
          description: Conta criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  bill: { $ref: '#/components/schemas/RecurringBill' }
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /recurring-bills/{cpf}/{billId}:
    put:
      tags: [Financeiro]
      summary: Atualizar conta recorrente
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
        - name: billId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecurringBillRequest' }
      responses:
        '200':
          description: Conta atualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '404':
          description: Conta não encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Financeiro]
      summary: Remover conta recorrente
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
        - name: billId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Conta removida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /recurring-bills/{cpf}/{billId}/pay:
    post:
      tags: [Financeiro]
      summary: Pagar conta recorrente (débito em conta ou cartão)
      description: |
        Pagamento manual de uma conta recorrente.

        - ACCOUNT_DEBIT debita users.balance (a transação PAYMENT aparece no
          extrato da home, nunca na fatura do cartão); saldo insuficiente
          retorna code=SALDO_INSUFICIENTE.
        - CREDIT_CARD consome credit_card_available_limit e grava PAYMENT com
          sufixo "(Faturado no Cartão)" (linha informativa na fatura aberta,
          sem alterar o total); limite insuficiente retorna code=LIMITE_INSUFICIENTE.
        - O ciclo da conta avança (next_billing_date +1 mês ou +1 ano) e o
          status volta a active com retryCount=0.
        - Upsert: se a conta não existir no banco (só no localStorage do
          frontend), registra com o billId enviado usando name/amount do body.
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
        - name: billId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecurringBillPayRequest' }
      responses:
        '200':
          description: Pagamento realizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    type: string
                    example: "Pagamento de Netflix realizado com sucesso via débito em conta."
                  bill: { $ref: '#/components/schemas/RecurringBill' }
                  transactionId:
                    type: string
                    example: "a1b2c3d4-0000-4000-8000-000000000000"
                  paymentMethod:
                    type: string
                    enum: [ACCOUNT_DEBIT, CREDIT_CARD]
                    example: ACCOUNT_DEBIT
                  newBalance:
                    type: number
                    format: float
                    description: Saldo atualizado (sempre retornado; para CREDIT_CARD é o saldo em conta, não o limite).
                    example: 6960.16
        '400':
          description: |
            Método inválido, saldo/limite insuficiente (code=SALDO_INSUFICIENTE |
            LIMITE_INSUFICIENTE) ou conta inexistente sem name/amount para upsert.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Acesso negado (não é o dono nem admin)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...